requirements.txt
pandas==2.2.3
numpy==1.26.4
matplotlib==3.10.1
statsmodels==0.14.4
scikit-learn==1.6.1
pmdarima==2.0.4
pyyaml==6.0.2
loguru==0.7.3
pydantic==2.11.3
torch==2.7.0
pytest-mock==3.14.0
pyspark==3.5.5
ray==2.48.0
darts==0.36.0
prophet==1.1.7
catboost==1.2.8
lightgbm==4.6.0
einops==0.8.1
mlflow==3.4.0
sqlalchemy==2.0.43
polars==1.33.1
pymysql==1.1.2
duckdb==1.4.0
